+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - User option to specify minimum width for a single breadcrumb entry ([#30](https://github.com/benji300/joplin-note-tabs/issues/30))
 - User option to specify the font size for the panel
 - User option to specify the background color when hovering over tabs
+- User option to display navigation buttons (backward/forward) below the tabs
 - Command to toggle panel visibility ([#16](https://github.com/benji300/joplin-note-tabs/issues/16))
 - Support to drag & drop notes from note list to pin as tabs ([#31](https://github.com/benji300/joplin-note-tabs/issues/31))
 - Support to drag & drop notes from [favorites plugin](https://github.com/benji300/joplin-favorites) to pin as tabs